requirements.txt
djangorestframework
djangorestframework-jwt==1.8.0
PyJWT
Django==2.2.2
pytz==2019.1
sqlparse==0.3.0
psycopg2==2.8.2